pyproject.toml
[tool.poetry]
name = "netvision"
version = "0.1.0"
description = ""
authors = ["BenderScript <6779302+BenderScript@users.noreply.github.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
streamlit = "^1.34.0"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
langchain = "^0.2.3"
langgraph = "^0.0.65"
langchain_openai = "^0.1.4"
langsmith = "^0.1.57"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
